+++ src/main/java/com/clanki/parser/Parser.java
@@ -3,9 +3,9 @@ package com.clanki.parser;
 import com.clanki.commands.AddCommand;
 import com.clanki.commands.Command;
 import com.clanki.commands.UnknownCommand;
-import com.clanki.objects.exceptions.EmptyFlashcardAnswerException;
-import com.clanki.objects.exceptions.EmptyFlashcardQuestionException;
-import com.clanki.objects.exceptions.InvalidAddFlashcardInputException;
+import com.clanki.exceptions.EmptyFlashcardAnswerException;
+import com.clanki.exceptions.EmptyFlashcardQuestionException;
+import com.clanki.exceptions.InvalidAddFlashcardInputException;
 
 public class Parser {
     public static final String QUESTION_START_INDICATOR = "/q";
