+++ src/test/java/com/clanki/objects/FlashcardTest.java
@@ -1,7 +1,8 @@
 
-package com.clanki.objects.flashcard;
+package com.clanki.objects;
 
 import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class FlashcardTest {
